NOBUG mod_surveypro: added template validation at application time too
Let's suppose the following scenario.
At time = 0 I save a usertemplate with a boolean item.
At time = 1 I upgrade my surveypro because the boolean plugin changed.
At time = 2 I go to apply my old usertemplate.
The content of the usertemplate is now out of date and its application may lead to an unexpected behaviour. A check of template integrity and of match between versions of the plugin used at save time and version of the current plugin is needed. The same applies for mastertemplate.