NOBUG mod_surveypro: changed visibility of methods and properties
In the frame of this same work I also gave attention to:

1 -----------------------
cover=0 as param in the URL of view.php left. It was 100% useless, but I needed many months to get it. Sorry! I am not higher than me.

2 -----------------------
Landing pages have now a speaking name. Names are now: << tabs >> _ << page >>.php like, for instance: view_search.php explaining that that file manages the page that comes out by selecting the "Layout" tab and choosing the "Search" page.
Landing pages of the "Survey" tab are named: view_*.php.
The same is true for the landing pages of master and user templates.
Corresponding class has been renamed with a name more similar to the calling page.

3 -----------------------
An awful error has been fixed. The story is here. Let's talk about a multipage surveypro. I am in page k. There I provide the answer to few items and I push the "<< back" button to return to page k-1. At submission time, surveypro code ALWAYS saves the answer for each item in page k. THIS IS TRUE EVEN IF THE PAGE WAS NOT 100% ANSWERED. This was requested by users a long time ago and now it is months that the module has this exciting feature. The problem I just fixed was with the date* (date, datetime, shordate...) items. When they were still set to "Choose..." the save process was rising up an issue. It has now been fixed.

4 -----------------------
$this->canxxx no longer exists. Capability are saved to local variables and recalculated at the beginning of each method.

5 -----------------------
view_preview has been created and has now the same dignity of the other landing pages.

6 -----------------------
Following the idea to reduce the visibility of methods and properties, I changed a bit the code of the landing pages in order to reduce the number of public properties.

7 -----------------------
bloody news: I see tabs.class.php in this branch. I am 100% sure it comes from a mistake of mines. I apologies.

8 -----------------------
Two errors were fixed in the usertemplates available in test/fixtures/usertemplates folder.
The story is here even if it is quite long. I divided it into four steps.
1) For "Select", "Radio button" and "Checkbox" items the ET (Editing Teacher or whoever is in charge of the creation of a surveypro) can create elements where value and label are different.
((1a) It is simple: In the "Options" field of the setup page of the "select" item write

IT::Italy
ES::Spain
FR::France
CH::Switzerland
DE::Germany

instead of (1b)

Italy
Spain
France
Switzerland
Germany
)
In detail: the ET can create a select like the following one:
Italy Spain France Switzerland Germany
where the <> "Spain" is different from the <<> "ES" that will be saved to the database.
Good.

2) For that items the ET can even choose a default. The vexed (it was the reason of some centuries of meetings) issue is: When the ET chooses a default what has to specify? The <> or the <>? I other words: if I want my html "select" default to Spain, what am I supposed to write in the "default" field of the setup page of the "select" item? "ES" or "Spain"? I was an herald of VALUE VALUE VALUE!! But at the end they won against me because...

3) because the ET can also use the "Other" field by defining the label of the "other" option of the select. That option will be bound to an input field like:
Italy
Spain
France
Switzerland
Germany
>_text" />

If the user chooses "Other (please specify)" he will be forced to enter a value in the Italy
Spain
France
Switzerland
Germany
>_text" value="myDefault" />

Well, the overall idea of this long story is that
a) the ET will ALWAYS write labels for each element of the "select"
b) it is NOT true he will ALWAYS write each value like explained in 3) where the content of the input text field is not set.

Because of this I changed, some months ago, the management of the default. It is no longer expecting VALUES but LABELS.

The issue I fixed is this: in the usertemplates provided in test/fixtures/usertemplates the defaultvalue provided was still a VALUE and not a LABEL.

9 -----------------------
dropped redirect from tab bar
The idea is that when admin/teacher selects a page in the tab bar, he knows what he expects.
If the survey is without items, a simple redirect when the admin/teacher asks for the preview page is not a good software answer because it can be misleading.
Because of this the code now provide a simple message suggesting the admin/teacher to go to add elements to the survey.
The same message for students is not supposed to never be seen because students are not allowed to ask for the preview.

10 ----------------------
Added utils class for very basic methods used everywhere

11 -----------------------
splitted the method to look after actions ('delete submission' and 'delete all submissions') in view.php between 'actions_execution' before $PAGE and 'actions_feedback' after $PAGE.
This was necessary because the content of the admin block depends on the number of submission in the survey.
If the survey has at least ONE submission the admin block does not display the usertemplate->apply item.
If the survey has NO submissions the admin block displays the usertemplate->apply item.
This is why I was forced to split the method.

Added utils class for very basic methods used everywhere

12 ----------------------
$cm is no longer passed to mod_surveypro_itembaseform as it is already present in $item that is passed as well

13 ----------------------
the property $context of the mod_surveypro_itembase class no longer exists. It too simple to calculate and quite always unused.

14 ----------------------
deleted defaults from properties and fixed some bug acording to no longer existing defaults.

15 ----------------------
Bloody last minute issue. Multipage survey. Respond. Go to page 2 where there is EACH item plugin. Do not fill but return back to page 1. Return now to page 2. Some defaults are missing. This is the last minute fix!!!!