NOBUG mod_surveypro: dropped redirect from tab bar
The idea is that when admin/teacher selects a page in the tab bar, he knows what he expects.
If the survey is without items, a simple redirect when the admin/teacher asks for the preview page is not a good software answer because it can be misleading.
Because of this the code now provide a simple message suggesting the admin/teacher to go to add elements to the survey.
The same message for students is not supposed to never be seen because students are not allowed to ask for the preview.