CONTRIB-6207 mod_surveypro: nicer message from import_csv() on error
At the moment, an incorrect definition of the parameters in the import form (view_import.php) can lead to an import failure. In this case the import process is aborted showing an error_message. This is not nice for teacher. I replaced it with a nicer $OUTPUT->notification

In the frme of this fix I also made minor changes to code to satisfy a bit codechecker.