NOBUG: Fix multilanguage test to ensure UI is in the expected lang.
Depending of browser language it may be possible for some tests
to be passing in some sites but not in others.

So I've created a new step:

   And I follow "languagename" in the language menu

In order to, always, ensure the desired language is being used.

Note it also includes the "Given remote langimport tests are enabled"
verification, so you'll need to add this to your config.php:

define('TOOL_LANGIMPORT_REMOTE_TESTS', true);

That's to avoid tests requiring external resources to be executed
by default (and your tests are using external resources).