CONTRIB-6268 mod_surveypro: added static styles per each item
With that addition I am able to apply a style to all the elements specified by pluginand item detail. For instance: date_select for the three drop down menu with day, month and year or date_chack for the noanswer checkbox.
This is specially useful for the autofill plugin that is changed in its aspect to look like a static text.