CONTRIB-6399 mod_surveypro: partial application of convert2static
If I am not wrong, PR#119 in github, is wrong and can not be merged because of the reasons I wrote in the discussion page of the PR.
In spite of this, part of the commit is fine and can be saved.
In this new commit I preserved the working part of convert2station.
This working part consists, mainly, in the deletion of the duplicated code.