NOBUG mod_surveypro: added explanation of use
Added explanation about the difference between $required and $minimumrequired and their compatibility for checkboxes elements.
This is needed because at first glace $minimumrequired greater than 0 may appear inappropriate for not required element.
The trick is in the semantic of the checkbox. In a word: student can never jump a checkbox element even if he/she leaves the alement untouched.
The added explanation describe exactly this.