CONTRIB-6467 mod_surveypro: new test for numbers using different formats
In the frame of this patch I found an undiscovered issue.
To submits a surveypro, students are redirected to the view_form page.
The view_form page verifies the submission, saves it and much more.
IF a submissions was really performed (no pause, no cancel, no previous page button was pressed) the thanks page is displayed, buttons to continue are displayed AND THE EXECUTION IS STOPPED.
This is very bad because if in that page the student switches the lang using the corresponding menu, he/she is redirected to the view_form page that is the one to submit one more response and not to the thanks page were he/she actually is.