CONTRIB-6476 mod_surveypro: improved submissions selection
Background.

Let's understand: submission === response.

At responses import time (from "moodle instance 1" to "moodle instance 2"),
the ownership of responses is assigned to the user with id == submission->userid.
It may happen that a submission has a userid not existing in "moodle instance 2".
In this case the ownership of the response is assigned to the user that is performing the import.
This user is not forced to be enrolled in the course so, at the end, it may happen that a submission in "moodle instance 2" is assigned to a not enrolled user.
This is the background.

The issue:
Usually the teacher don't want to see submissions from user not enrolled.
At the same time the admin WANTS to see its submissions even if he is not enrolled in the course.
So... which submission has to be selected by queries?

The agreed answer is this:

     * Teachers is the role of users usually accessing reports.
     * They are "teachers" so they care about "students" and nothing more.
     * If, at import time, some records go under the admin ownership
     * the teacher is not supposed to see them because admin is not a student.
     * In this case, if the teacher wants to see submissions owned by admin, HE HAS TO ENROLL ADMIN with some role.
     *
     * Different is the story for the admin.
     * If an admin wants to make a report, he will see EACH RESPONSE SUBMITTED
     * without care to the role of the owner of the submission.

This issue adds this feature to the module.