CONTRIB-6486 mod_surveypro: added jumper to delayedusers report
If a surveypro activity is divided into groups
reports should provide a jumper between groups
in order to facilitate the access to information.

I am sorry, I was not able to add this new feature without changing the code into attachment too.
It was supposed to be over but I was forced to change the chain of calls. I apologize.