CONTRIB-6542 mod_surveypro: enforced if condition
Create a multipage surveypro.
Put a non mandatory checkbox element in the first page and something else in the second one.
Go to fill the surveypro.
Leave each checkbox of the element in the first page unselected and push the `Next page >>` button.
Return to the first page with `<< Previous page` and look at the checkbox element.
No answer is selected by mistake.