Removed useless if
A long time ago for surveypros spanning more pages when the user paused a submission the current page was not saved.
In that situation the surveys with only one page were forced to neglect the allowpauseresume option.
Now the current page is actually saved so even if the survey lies on a single page the "Pause" button should be displayed.