added code to drop unexpected reload
After the item deletion action, if the user reload the page, the deletion is performed again rising up an error.
If the item to drop is not in the db, this means that the user already deleted it and is reloading the page.
In this case, stop the deletion execution.