added method to get static content for formats elements
Pagebreak and fieldsetend don't use content at all but they must be described with some sort of content when the list of item of a layout is requested. Because of this I added a method passing the the "static" content to the methods supposed to display the list of items.