Element label is finally written as it is
Now the user if free to add a picture in the element label and to display it in the form even if the position of the content is on the left of the item.
From now on the user is responsible of what he write in the question content. I no longer drop tags from his input.
The height of the rows in the list of items will be corrected, as it should be, using css.