Fixed the way to filter what student submits
With the introduction of interactive fieldset those format elements are now equipped with open/close triangle and they submit their own state. Because of this the paused survey tries to save the information submitted by interactive fieldset. This causes the "Call to undefined method mod_surveypro_utility_layout::delete_answer()"