Changed standing report query
The idea is: if the user is allowed to pause/resume the survey he has to be counted among late users if he didn't submit its answer OR if he submitted only IN PROGRESS responses.
Further more... if the required conditions are defined as "The user has to sumbit at least x responses" the user has to be considered "late user" if he submitted AS CLOSED less than x responses.