Dropped "return true" to keep the button enabled
At the end of the export to excel of saved responses the "Continue" form button was unexpectedly disabled.
This is an issue if people ask to download to excel twice (for instance for "closed" and "in progress" responses separately).
It seems that dropping "return true" or changing it with exit; fixes the issue.