Lower the Maximum file size to just 2MB in scenario
The php default is 2MB, unless upload_max_filesize is modified.

In order to allow the test to pass in any environment, like
GHA, for example, that runs with php default values, we are going
to lower the test to use those 2MB that should be ok for any site.