Fixed the routine to drop no longer allowed answers
The context of this fix is this: let's suppose I have a surveypro with parent child relation spanning multiple  pages.
On page 1 I have a branching item.
On page 2 I have the child of the branching item AND a question without parent.
I fill the surveypro.
I add an answer for the parent item and I move to page 2.
Here I add an answer for the child item and the free item.
At the end I return back to page 1.
I change my answer to the parent item.
I return to page 2, I fill the free item and I submit.
The submission HAS TO BE CLOSED and not marked as "in progress" because some item has answer not verified.

The original answer to the child item should, actually, be dropped when the student returns to page 2 for the seconf time.
This path corrects the routine to drop answers to this kind of items.