Changed behaviour for read-only responses buttons
Browsing a submitted response in read-only mode (by pushing the glasses icon in the responses list page) the student is allowed to see each page of the surveypro even if they werte not allowed at submission time.
With this patch I changed the way nextpage is calculated and I used it for the buttons to browse the read-only responses.

Each behat test runs fine.
This PR has been built on top of #707.