Added custom number to items displayed in full width
For question displayed with the text spanning all the width of the form the question number was not always added.
The reason is that the routine expects the question starting wit <p> but if it starts with <p class"some_class_name"> the routine does not work properly.

This patch was written on top of next_allowed_page because otherwise the file classes/local/form/userform.php does not exist.