Added check for first page non allowed
Let's suppose a surveypro having only reverved items in the first page.
When the simple user starts to fill the form, he/she can not go to the first page of the surveypro BY DEFAULT but there must be a check for the first available page.
This routine adds the described check.

This PR has been created on top of #710.