Fixed the routine to drop undesiderd answers (rev 04)
The context of this fix is this: let's suppose I have a surveypro with parent child relation spanning multiple pages.

Simple scenario
On page 1 I have a branching item.
On page 2 I have the child of the branching item AND a question without parent.
I fill the surveypro.
I add an answer for the parent item and I move to page 2.
Here I add an answer for the child item and for the free item.
At the end I return back to page 1.
I change my mind and the answer to the parent item.
I return to page 2, I fill the free item and I submit.
The submission HAS TO BE CLOSED and not marked as "in progress" because some item has answer not verified.

The original answer to the child item should, actually, be dropped.
This patch corrects the routine to drop answers to this kind of items.

Advanced scenario
On page 1 I have a branching item.
On page 2 I have an item depending from branching item when answer is 1.
On page 3 I have an item depending from branching item when answer is 0, one more item depending from branching item when answer is 1 AND one free item (without dependencies).
I fill the surveypro.
I answer 1 to the parent item and I move forward.
I find an item in page 2 and I provide an answer.
I move forward and, in page 3, I find the item depending from branching item when answer is 1 AND the free item.
I provide an answer to both of them and I move backward.
I land in page 2. I am free to leave the item untouched or to change the answer.
I move backward once more and I change (in page 1) the answer to branching item (from 1 to 0) and I move forward.
I land in page 3 and I see only the item depending from branching item when answer is 0 and the free item (without dependencies).
I fill the page and I submit.
The submission HAS TO BE CLOSED and not marked as "in progress" because some item has answer not verified.