Fixed wrong default for surveyprofield_multiselect
The issue, IMHO, is here since the very first release of surveypro.
Asking for the list of items (as teacher or admin, of course) the red dot icon allow the user to mark the item (the question corresponding to the item) as mandatory or not.
Of course, if the default of the item is "No answer", that item can NOT be set to mandatory. (if it is mandatory you can not answer: "I don't answer"!)
The user is supposed to change the default to a different value and ONLY AFTER he is allowed to set the item to mandatory.
Asking for the list of items, items type = 'multiselect' were equipped with the possibility to change them to mandatory or not mandatory WITHOUT CARE to the default. This was due to the routine "item_canbemandatory" in itembase.php that was not taking care of the way some item are marked as mandatory.
Actually there are two kind of items:
 1) those for which (like the boolean)
     defaultoption discriminates on the desired type of default: "Custom", "Invite", "No response"
     and, if defaultoption == "Custom", defaultvalue intervenes and declares which custom default is chosen.

 2) those for which (like multiselect).
     noanswerdefault = 1 means "No response".

I corrected the code and, in the frame of the correction, I found that the boolean field surveyprofield_multiselect.noanswerdefault was set to have 2 as default. Of course this is a nonsense.