Modified separators definition
Now vertical radio button with the option "other" and the field "specify" close the option "other" are correctly displayed.
The "specify" field had a distance from the corresponding button proportional to the indent. Now this is no longer true.
In the frame of this PR I also copied the same solution (for the distance between checkbox and the "specify" field) to the checkbox item that was affected by the same problem.