Bump to m402
This PR changes the moodle required version and the surveypro version.

This PR fails because it is missing the declaration of core_external\external_api; in all the files stored into classes/local/ipe.
Because of this I deleted the specific PR written to add those declarations and I included them here.

The feature in surveypro/field/fileupload/tests/behat/submit_attachment.feature was failing too. The failure was due to the new editor that saves the content of thextare always crafting it between html tags. Because of this I added the user preferences to the scenario and all started to work fine once again.

This is because in this PR that was supposed to only increase required moodle version and surveypro version was included some more modified code.