Added a routine to get separator
Radio button questions with only two options are, in my mind, boolean questions.
This is not true in the mind of some teacher using surveypro.
If you create a surveypro with a radiobutton item with the simple question: Who do you love the most?
with "Choose..." at the beginning (default = invitation), mandatory, and with only two options: "mum" and "dad", surveypro crashes.
This PR fixes this issue due to wrong definition of the separator between item elements.