fixed number validation master
This PR fixes two errors regarding numeric sub-plugin.
Create an item type numeric providing, for instance, a lowerbound.
Edit the item and try to wipe the lowerbound.
It is not deleted.

Going to fill a form providing 0,0 (in italian) as number the number is wrongly classified as not a number.