numeric_validation_error_m402
This PR fixes two errors regarding numeric sub-plugin.
Crate an item type numeric providing, for instance, a lowerbound.
Edit the item and try to wipe the lowerbound. It is not deleted.

Going to fill a form providing 0,0 (in italian) as number the number is wrongly classified as not a number.