Added userform_get_separator to checkboxes too.
For consistency with what was done for radiobutton type items, I added the userform_get_separator method to checkbox type items, too.
Unlike what I did for radiobutton items, however, I cannot add a unit test here because it would fail.
The error is described in the  comment of the method itself.