Modified wrong links
Using theme classic the navigation between pages of the module (Surveypro, Layout, Reports, Tools, uTemplate and mTemplate) are driven by the navigation menu. In that menu the link for Reports, uTemplate and mTemplate are wrong.
I unifield the source for the url so, from now on, they will always be equal between navigation and admin block.
This PR should be cherry-picked to MOODLE_401|402|403_STABLE too.