Changed the mform element to display autofill
The issue is this: giving a parent to an autofill item the autofill form element is disabled at first form load (because it is an autofill and the <input type="text"> has disabled="disabled") but changing the answer in the parent item the autofill may be enabled by js.
Now I changed the mform element. It is no longer a disabled text input but a plain label.
In the frame of this PR I also changed the name of method "userform_set_prefill" to the more correct "userform_get_prefill".

Changed the mform element to display autofill

The issue is this: giving a parent to an autofill item the autofill form element is disabled at first form load (because it is an autofill and the <input type="text"> has disabled="disabled") but changing the answer in the parent item the autofill may be enabled by js.
Now I changed the mform element. It is no longer a disabled text input but a plain label.
In the frame of this PR I also changed the name of method "userform_set_prefill" to the more correct "userform_get_prefill".