Format items and autofill no longer can be child items
I always postponed this intervention saying "parent child relation is useless for those items and none will never use it" but now I found an issue in autofill items that make it convenient, from the development point of view, to fix this issue now.
The issue in autofill items is fixwd in the next PR. I guess it will #914.

Format items and autofill no longer can be child items

I always postponed this intervention saying "parent child relation is useless for those items and none will never use it" but now I found an issue in autofill items that make it convenient, from the development point of view, to fix this issue now.
The issue in autofill items is fixwd in the next PR. I guess it will #914.